autopraesi_python/requirements.txt
python-pptx
openpyxl
requests
lxml
beautifulsoup4
